# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Gestion Compte API",
        "description": "API pour la gestion des comptes bancaires",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Serveur local"
        },
        {
            "url": "https://gestion-compte-amycolle.onrender.com",
            "description": "Serveur de production"
        }
    ],
    "paths": {
...
        "/ndiaye/api/v1/comptes": {
            "get": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Lister les comptes",
                "description": "Récupère une liste paginée des comptes bancaires",
                "operationId": "listComptes",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Nombre d'éléments par page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 5,
                            "maximum": 100,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Numéro de la page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Type de compte (courant, epargne, cheque)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "courant",
                                "epargne",
                                "cheque"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des comptes récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Créer un nouveau compte",
                "description": "Crée un nouveau compte bancaire avec les informations du client",
                "operationId": "createCompte",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Compte créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Données invalides",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Les données fournies sont invalides"
                                        },
                                        "error": {
                                            "properties": {
                                                "code": {
                                                    "type": "string",
                                                    "example": "VALIDATION_ERROR"
                                                },
                                                "details": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "SERVER_ERROR"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/ndiaye/api/v1/comptes/{numero}": {
            "get": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Rechercher un compte par numéro",
                "description": "Récupère les détails d'un compte bancaire spécifique par son numéro",
                "operationId": "getCompteByNumero",
                "parameters": [
                    {
                        "name": "numero",
                        "in": "path",
                        "description": "Numéro du compte bancaire",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Compte trouvé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Supprimer un compte (soft delete)",
                "description": "Marque un compte bancaire comme supprimé sans le supprimer définitivement",
                "operationId": "deleteCompte",
                "parameters": [
                    {
                        "name": "numero",
                        "in": "path",
                        "description": "Numéro du compte bancaire à supprimer",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Compte supprimé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Compte supprimé (soft delete) avec succès"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Compte non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Aucun compte trouvé"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "SERVER_ERROR"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/ndiaye/api/v1/comptes/telephone/{telephone}": {
            "get": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Rechercher un compte par téléphone",
                "description": "Récupère les détails d'un compte bancaire par le numéro de téléphone du client",
                "operationId": "getCompteByTelephone",
                "parameters": [
                    {
                        "name": "telephone",
                        "in": "path",
                        "description": "Numéro de téléphone du client",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Compte trouvé avec succès"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}